EP-3079 Refactor "layer catalog" functionality into backend class
- depends on related refactor in openeo-python-driver
- this replaces `getImageCollection` with `GeoPySparkLayerCatalog.load_collection`
- note: moving kerberos to utils was to avoid circular import issues #12
